reorganize project  architecture
