+++ .gitignore
@@ -1,2 +1,3 @@
 scheduler.*
+scheduler2.db
 .vscode
